move psql config
