+++ infrastructure/psql_client.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/s14t284/apple-maitained-bot/config"
+
 	"github.com/labstack/gommon/log"
 	"github.com/s14t284/apple-maitained-bot/domain/model"
 	"gorm.io/driver/postgres"
@@ -15,16 +17,7 @@ type SQLClient struct {
 	Client *gorm.DB
 }
 
-// PsqlConfig psqlに接続するための設定値
-type PsqlConfig struct {
-	Host     string
-	UserName string
-	Password string
-	Port     int
-	Database string
-}
-
-func createDataSourceName(config PsqlConfig) string {
+func createDataSourceName(config config.PsqlConfig) string {
 	return fmt.Sprintf("user=%s password=%s host=%s port=%d dbname=%s sslmode=require TimeZone=Asia/Tokyo",
 		config.UserName, config.Password, config.Host, config.Port, config.Database)
 }
@@ -43,12 +36,12 @@ func initializeTables(dbClient *gorm.DB) {
 }
 
 // PsqlNewClientImpl psqlに接続したgormクライアントを返却
-func PsqlNewClientImpl(config PsqlConfig) (*SQLClient, error) {
+func PsqlNewClientImpl(c config.PsqlConfig) (*SQLClient, error) {
 	// deploy先では環境変数から接続先情報を取得できる
 	dataSourceName := os.Getenv("DATABASE_URL")
 	// localでは予め .env に入力しておいた情報から接続先を特定
 	if dataSourceName == "" {
-		dataSourceName = createDataSourceName(config)
+		dataSourceName = createDataSourceName(c)
 	}
 	dbClient, err := gorm.Open(postgres.Open(dataSourceName), &gorm.Config{})
 	if err != nil {
